refactor(graph-merge): extract the identity survivor rule and semantic key into identity-three-way
